+++ crates/codegen/CHANGELOG.md
@@ -8,6 +8,8 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+## [2.0.0] - 2025-08-28
+
 ### Added
 
 - `Generator` now has a `producer_version` method so the version in the